Fix radial gauge on mobile

+++ lib/approverlist_funder_screen.dart
@@ -351,7 +351,7 @@ class DetailScreenApprovedList extends StatelessWidget {
                                         .toString()),
                               ),
                             ),
-                            _getRadialGauge( snapshot.data[index]['Creditscore'])],
+                            _getRadialGauge( snapshot.data[index]['Creditscore'].toDouble())],
                         ),
                       );
                     });
