+++ lib/approverlist_screen.dart
@@ -435,7 +435,7 @@ class DetailScreenApprovedList extends StatelessWidget {
 
                                     ),
                                   ),
-                                  _getRadialGauge( snapshot.data[index]['Creditscore'])],
+                                  _getRadialGauge( snapshot.data[index]['Creditscore'].toDouble() )],
                               ),
 
                             );
